+++ client/wwwroot/js/src_portfolio-os/portfolio-os_auth.header.js
@@ -1,0 +1,2 @@
+﻿// %import<<GH 'js-cookie/js-cookie/master/src/js.cookie.js'
+// @import 'portfolio-os_auth'